Accept capitalized Project in metadata
Files with `type: Project` currently don't show up as projects because of the capitalization.  This fix updates the regex to accept both `type: project` and `type: Project`.  See https://github.com/Thinkful-Ed/curric-node-001-v4/issues/16 for an example issue which would be fixed by this without needing updates to the metadata.